requirements.txt
fastapi>=0.115
httpx>=0.27
uvicorn>=0.30
pytest>=8.0
PyJWT>=2.8
python-multipart>=0.0.9
